Exchanger: add handle error by confirm
