+++ src/components/error/Error.css
@@ -1,0 +1,12 @@
+.error {
+    position: absolute;
+    top: 25%;
+    left: 35%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    padding: 20px;
+    border-radius: 10px;
+    background-color: var(--gray);
+    color: var(--bg);
+}
